Extend syntax for comprehensive highlighting
This commit extends the syntax to cover most aspects of the language,
including highlighting of illegal uses of keywords and operators. It can
highlight nested generic types, type literals, and block parameters.
Standard types, control structures and annotations are now placed in the
'support' scope, and other types and operators are in the 'storage'
scope.

The redundant A-Za-z regex has been converted to A-z in many places.