Track open braces for better interpolation
The interpolation matches against a closing brace, and the only way to
avoid having other constructs' closing braces close the interpolation is
to match every brace. Constructs which consumed their opening brace now
stop right before the brace, and all braces which are not consumed
normally are matched together. This means that highlighting of block
parameters is no longer possible.

The syntax for identifying opening generic parameters has also been
tidied a little to try and ignore non-spaced binary operators.