Add simple block parameter highlighting
Correctly identifying all block parameters is incredibly difficult
because it essentially requires discovering that what has already been
highlighted were supposed to be parameters and backtracking. This change
adds support for highlighting simple block parameters, where the
accompanying arrow is on the same line and there are no closing braces
in the parameter list. This is the standard arrangement for most block
parameters, so it should suffice. Crucially, there shouldn't be any
false positives against correct Grace code (better to fail to highlight
than destroy other parts of the highlighting) and it shouldn't interfere
with brace matching for string interpolation.