Add generic parameters to method signatures
Generic arguments were being highlighted, but not generic parameters on
method definitions. Now definitions of non-operator methods will
highlight generic delimiters as keywords and identifiers inside the
delimiters as types (regardless of their casing). This change also
includes more specific highlighting of operator method definitions,
including highlighting the prefix keyword for prefix methods (prefix is
only a keyword in this context) and labelling mixing of names and
operators as illegal.