Prevent outer brace from capturing block params
The beginning match for block parameters searched from an opening brace
to the arrow through anything which was not a closing brace. Because the
matching is done left-to-right, and did not require actually capturing
the open brace, any open brace (including braces for string
interpolation) could capture the block parameters, consuming the actual
inner open brace for the block. Once the real open brace was consumed,
its closing brace was not matched and would be erroneously matched
against the outer open brace. Now the beginning match cannot pass
through either opening or closing braces. This remains an imperfect
solution (as braces may appear in the parameter list), but this match
must be performed by a regular expression and the syntax for the
parameter list is not regular.

This change also adds illegal highlighting of keywords and operators in
the parameter list.