Fix categorisation of untyped method parameters
The highlighting for method parameters only matched if the parameter was
typed or was the last parameter in the list. Now parameters are also
identified by appearing before a comma.

This change also moves method highlighting to use the function
categorisation instead of method, as this is more widely recognised as a
highlighting category.