Merge method and signature parsing for operators
The highlighting of signatures only addressed simple method names, and
did not include prefix and binary operator definitions. These were
handled just for method declarations, and not for signatures in type
declarations. This change moves the handling of the operators into the
signature highlighting, with method declarations already handled by the
remaining rule.

The highlighting of signatures has also been tidied up. Erroneous use of
assignment and general operators are reported more accurately, and
failure to match the end of potentially multi-line expressions which are
also ended by a newline should occur less often.