Re-license this software to the MIT License
This bundle is used (improperly) by several different pieces of software
under the terms of the GPL. As the priority for this software is to see
Grace highlighted correctly on as many compatible platforms as possible,
it is being re-licensed under the MIT License. As the sole contributor
to this software, I have the appropriate permissions to make this
change.